Updated Keyboard information and created BOM
The major update is me completing the Bill of Materials (BOM) for the keyboard. I have since ordered all the parts listed, but I was having a lot of trouble finding 5mm thick acrylic large enough (in the US) to cut the pieces from. There are a lot of European suppliers who would have done it on the cheap, and there were a lot of US suppliers supplying US-based thicknesses (3/8 inch, not 5mm). So, I ended up having to spend about $30 on the acrylic from Amazon in small pieces, so I generated 2 new dxf files to represent the separate halves of keyboard. In addition, I made the decision to purchase Zilent switches and SA Ice Cap keycaps, so I'm really high rolling now.